raspberry_pi/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
websockets==12.0
picamera2==0.3.16
opencv-python==4.8.1.78
numpy==1.24.3
RPi.GPIO==0.7.1
